script/scrap_website/requirements.txt
pymongo==4.6.2
selenium==4.18.1
python-dateutil
webdriver-manager==4.0.0